Fix issue #103, ternary operator result must have the same type.

+++ uicdm/src/main/java/ucar/nc2/ui/gis/shapefile/DbaseData.java
@@ -96,7 +96,7 @@ public class DbaseData {
         break;
       case 'F': /* binary floating point */
         if (desc.FieldLength == 4) {
-          numeric[n] = (double) Swap.swapFloat(field, 0);
+          numeric[n] = Swap.swapFloat(field, 0);
         } else {
           numeric[n] = Swap.swapDouble(field, 0);
         }
